api/requirements.txt
requests
openai
azure-functions
